+++ src/openhems/modules/network/node.py
@@ -175,7 +175,9 @@ class ApplianceConstraints():
 class OnNodeManager:
 	"""
 	Manage node on/off state
-	The goal is to detect if the user manually switch on/off the node and to avoid to switch on/off too often.
+
+	The goal is to detect if the user manually switch on/off
+	  the node and to avoid to switch on/off too often.
 	"""
 	def __init__(self, node, feeder:Feeder):
 		self._node = node
@@ -401,8 +403,8 @@ class Node:
 		return currentPower
 
 	def getMaxPower(self):
-		"""
-		Get max power 
+		""" 
+		Get max power in Kwh
 		"""
 		return self._maxPower.getValue()