+++ src/openhems/modules/util/__init__.py
@@ -11,3 +11,6 @@ from .notification_manager import NotificationManager, MessageHistory
 from .project_configuration import ProjectConfiguration
 from .recorder import Recorder
 from .logging import getLogger, filer, get_log_file_path
+from .error import DangerousStateException
+from .utility import obj_differ
+from .json import json_default