+++ src/openhems/modules/web/pages/02_Configuration.py
@@ -7,6 +7,7 @@ Page for OpenHEMS configuration, with two modes:
 import time
 import sys
 import re
+import os
 import json
 from pathlib import Path
 from enum import Enum
@@ -15,7 +16,9 @@ import yaml
 import jsonschema
 import streamlit as st # pylint: disable=E0401
 from streamlit_monaco_yaml import monaco_editor # pylint: disable=E0401
-from openhems.modules.util.json import json_default
+from openhems.modules.util import (
+    json_default, obj_differ
+)
 from openhems.modules.web.web_streamlit import get_logger
 
 # pylint: disable=wrong-import-position
@@ -554,7 +557,8 @@ Vous pourrez ensuite modifier les paramètres finement dans l'interface standard
             st.rerun()
     with col2:
         if st.button("❌ Annuler", use_container_width=True):
-            st.stop()
+            st.session_state.config_ui_editor = ConfigEditionState.YAML_EDITOR.value
+            st.rerun()
 
 def basic_configure_battery():
     """
@@ -648,14 +652,25 @@ def basic_configure(config_page, cancel=False):
 
     return state
 
-def yaml_editor_page(config_page):
+def yaml_editor_page(config_page, conf):
     """
     Display the YAML editor page (as plaintext).
     """
     st.title("✍️ Éditeur YAML Avancé")
     json_schema = load_schema()
     state = ConfigEditionState.YAML_EDITOR.value
-    with open(config_page, "r", encoding="utf-8") as f1:
+    use_tmp_file = (not st.session_state.conf_has_updated_running) and st.session_state.conf_differs
+    if use_tmp_file:
+        # We will edit a temporary file with the running conf (Can be many files combined)
+        # We will swap them at end
+        config_page_edit = str(config_page.parents[0]) + "/.tmp." + str(config_page.name)
+        # print("config_page_edit:", config_page_edit, "; from:", config_page)
+        conf = get_current_configuration()
+        with open(config_page_edit, "w", encoding="utf-8") as f0:
+            f0.write(yaml.dump(conf))
+    else:
+        config_page_edit = config_page
+    with open(config_page_edit, "r", encoding="utf-8") as f1:
         initial_text = f1.read()
         # st.write(f"ddd{initial_text}.")
         monaco_return = monaco_editor(
@@ -675,7 +690,12 @@ def yaml_editor_page(config_page):
             col1, col2 = st.columns(2)
             with col1:
                 if st.button("💾 Sauvegarder"):
-                    save_config(config=yaml_content, config_page=config_page, schema=json_schema)
+                    save_config(config=yaml_content,
+                                config_page=config_page_edit,
+                                schema=json_schema)
+                    if use_tmp_file:
+                        os.rename(config_page_edit, config_page)
+                        st.session_state.conf_has_updated_running = True # (even if no change)
                     st.success("✅ Fichier YAML sauvegardé !")
             with col2:
                 if st.button("✏️ Assistant d'édition de configuration"):
@@ -684,48 +704,72 @@ def yaml_editor_page(config_page):
             st.error(f"❌ Fichier YAML invalide selon le schéma JSON. : {e}")
     return state
 
+# @st.cache_data(ttl=3600)
+def get_current_configuration():
+    """
+    return the current configuration, the running one if configurator_path has not been edited,
+    configurator_path else. Set 'conf_has_updated_running'.
+    """
+    configurator_path = st.session_state.configurator_path
+    if st.session_state.conf_has_updated_running:
+        print("get_current_configuration(",configurator_path,")")
+        with open(configurator_path, "r", encoding="utf-8") as f1:
+            return yaml.safe_load(f1)
+    else:
+        conf0 = st.session_state.configurator.retrieveYamlConfig()
+        with open(configurator_path, "r", encoding="utf-8") as f1:
+            conf1 = yaml.safe_load(f1)
+        if obj_differ(conf0, conf1):
+            st.session_state.conf_differs = True
+        else:
+            st.session_state.conf_differs = False
+        return conf0
+    return {}
+
 def configure_page():
     """
     Main function to display the configuration page.
     It manages the state of the page (YAML editor or assistant).
     """
     # Manage the page
+    # pylint: disable=too-many-branches
     if 'configurator_path' not in st.session_state:
         OpenhemsHTTPServer.init_session()
     configurator_path = st.session_state.configurator_path
-    with open(configurator_path, "r", encoding="utf-8") as f1:
-        conf = yaml.safe_load(f1)
-        if conf is None:
-            conf = {}
-        has_nodes = len(conf.get("network", {}).get("nodes", [])) != 0
-        if "config_ui_editor" in st.session_state:
-            edition_state = st.session_state.config_ui_editor
-        elif not has_nodes:
-            edition_state = ConfigEditionState.ASSISTANT.value
-        else:
-            edition_state = ConfigEditionState.YAML_EDITOR.value
-        new_state = edition_state
-        if edition_state==ConfigEditionState.ASSISTANT.value:
-            new_state = basic_configure(configurator_path, cancel=has_nodes)
-        elif edition_state==ConfigEditionState.ASSISTANT_WARNING.value:
-            st.warning(
-                "**⚠️ Risque de perte d’informations ou de formatage**\n\n"
-                "L’assistant d'édition est imparfait et peut occasionner"
-                " des pertes dans votre configuration "
-                "a minima dans la mise en page si vous l'avez édité manuellement."
-                " Souhaitez-vous continuer ?"
-            )
-            col1, col2 = st.columns(2)
-            with col1:
-                if st.button("✅ Oui, ouvrir l’assistant"):
-                    new_state = ConfigEditionState.ASSISTANT.value
-            with col2:
-                if st.button("❌ Non, annuler"):
-                    new_state = ConfigEditionState.YAML_EDITOR.value
-        elif edition_state==ConfigEditionState.YAML_EDITOR.value:
-            new_state = yaml_editor_page(configurator_path)
-        if new_state!=edition_state:
-            st.session_state.config_ui_editor = new_state
-            st.rerun()
+    conf = get_current_configuration()
+    if st.session_state.conf_has_updated_running:
+        st.warning("⚠️ La configuration a été modifiée,\n " \
+            "pour la prendre en compte veuillez redémarrer"
+        )
+    has_nodes = len(conf.get("network", {}).get("nodes", [])) != 0
+    if "config_ui_editor" in st.session_state:
+        edition_state = st.session_state.config_ui_editor
+    elif not has_nodes:
+        edition_state = ConfigEditionState.ASSISTANT.value
+    else:
+        edition_state = ConfigEditionState.YAML_EDITOR.value
+    new_state = edition_state
+    if edition_state==ConfigEditionState.ASSISTANT.value:
+        new_state = basic_configure(configurator_path, cancel=has_nodes)
+    elif edition_state==ConfigEditionState.ASSISTANT_WARNING.value:
+        st.warning(
+            "**⚠️ Risque de perte d’informations ou de formatage**\n\n"
+            "L’assistant d'édition est imparfait et peut occasionner"
+            " des pertes dans votre configuration "
+            "a minima dans la mise en page si vous l'avez édité manuellement."
+            " Souhaitez-vous continuer ?"
+        )
+        col1, col2 = st.columns(2)
+        with col1:
+            if st.button("✅ Oui, ouvrir l’assistant"):
+                new_state = ConfigEditionState.ASSISTANT.value
+        with col2:
+            if st.button("❌ Non, annuler"):
+                new_state = ConfigEditionState.YAML_EDITOR.value
+    elif edition_state==ConfigEditionState.YAML_EDITOR.value:
+        new_state = yaml_editor_page(configurator_path, conf)
+    if new_state!=edition_state:
+        st.session_state.config_ui_editor = new_state
+        st.rerun()
 
 configure_page()