+++ src/openhems/modules/web/pages/04_About.py
@@ -18,8 +18,10 @@ def about_page():
     st.markdown(f"""
     **{conf.getName()}** version {conf.getVersion()}
     Licence : {conf.getLicence()}
-    [Site web]({conf.getUrls().get('Homepage', '#')}) | [Documentation]({conf.getUrls().get('Documentation', '#')})
-    [Source code]({conf.getUrls().get('Repository', '#')}) | [Bug tracker]({conf.getUrls().get('Issues', '#')})
+    [Site web]({conf.getUrls().get('Homepage', '#')}) |
+    [Documentation]({conf.getUrls().get('Documentation', '#')})
+    [Source code]({conf.getUrls().get('Repository', '#')}) |
+    [Bug tracker]({conf.getUrls().get('Issues', '#')})
     Contact : {conf.getContact()}
 
     N'hésitez pas à contribuer au projet sur [GitHub]({conf.getUrls().get('Source', '#')})